Adds a separate tab to display maps in details activity
