+++ app/build.gradle
@@ -24,25 +24,25 @@ dependencies {
 
     testCompile 'junit:junit:4.12'
 
-    compile 'com.google.android.gms:play-services:9.4.0'
-    compile 'com.jakewharton:butterknife:6.1.0'
 
-    compile 'com.google.android.gms:play-services-maps:9.4.0'
-    compile 'com.android.support:design:24.2.0'
     compile files('libs/apache-commons-lang.jar')
     compile files('libs/gson-2.2.4.jar')
 
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:24.2.0'
     compile 'com.android.support:design:24.2.0'
 
+
+
+
+    compile 'com.google.android.gms:play-services:9.4.0'
+    compile 'com.jakewharton:butterknife:6.1.0'
+    compile 'com.google.android.gms:play-services-maps:9.4.0'
+    compile 'com.android.support:design:24.2.0'
+    compile 'com.android.support:appcompat-v7:24.2.0'
     compile 'io.reactivex:rxandroid:1.2.1'
     compile 'com.squareup.retrofit2:retrofit:2.1.0'
     compile 'com.squareup.retrofit2:adapter-rxjava:2.1.0'
     compile 'com.squareup.retrofit2:converter-gson:2.1.0'
     compile 'com.google.code.gson:gson:2.4'
     compile 'io.reactivex:rxjava:1.1.7'
-
-
-
 }
