+++ app/src/main/java/com/festember16/app/DetailsFragment.java
@@ -49,28 +49,17 @@ import butterknife.InjectView;
 /**
  * Created by vishnu on 8/9/16.
  */
-public class DetailsFragment extends Fragment implements OnMapReadyCallback {
+public class DetailsFragment extends Fragment {
 
 
     public static final String ID = "ID";
-    public static final int PERMISSION_REQUEST_CODE = 1000;
-    public static final int LOCATION_ENABLE_REQUEST_CODE = 1002;
     public static final String LOG_TAG = "DetailsFragment";
 
-    private static final String[] PERMISSIONS = {
-            android.Manifest.permission.ACCESS_COARSE_LOCATION,
-            android.Manifest.permission.ACCESS_FINE_LOCATION
-    };
+
     public static final String REGISTERED_EVENTS = "REGISTERED EVENTS";
 
-
-    public GoogleMap mMap;
-    SupportMapFragment mapFragment;
-    private FragmentActivity myContext;
-    private boolean isLocationEnabled = true;
-    private boolean isPermissionGiven = true;
     private boolean hasRegistered;
-    Events events;
+    public static Events events;
 
     @InjectView(R.id.eventVenue)
     TextView eventVenue;
@@ -78,8 +67,8 @@ public class DetailsFragment extends Fragment implements OnMapReadyCallback {
     TextView eventDesc;
     @InjectView(R.id.eventStartTime)
     TextView eventStartTime;
-    @InjectView(R.id.eventCluster)
-    TextView eventCluster;
+//    @InjectView(R.id.eventCluster)
+//    TextView eventCluster;
     @InjectView(R.id.eventPoster)
     ImageView eventPoster;
     @InjectView(R.id.registerButton)
@@ -133,26 +122,9 @@ public class DetailsFragment extends Fragment implements OnMapReadyCallback {
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        FragmentManager fm = getChildFragmentManager();
-        mapFragment = (SupportMapFragment)
-                fm.findFragmentById(R.id.mapPreview);
-
-        if(mapFragment == null){
-            mapFragment = SupportMapFragment.newInstance();
-            fm.beginTransaction().replace(R.id.mapContainer, mapFragment);
-        }
-
-        ViewGroup.LayoutParams layoutParams = mapFragment.getView().getLayoutParams();
-
-        DisplayMetrics displayMetrics = new DisplayMetrics();
-        getActivity().getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
-
-        layoutParams.height = (int) (displayMetrics.heightPixels * 0.3);
-
-        mapFragment.getView().setLayoutParams(layoutParams);
         eventVenue.setText(events.getVenue());
         eventStartTime.setText(events.getStartTime());
-        eventCluster.setText(events.getCluster());
+       // eventCluster.setText(events.getCluster());
         //Todo: set eventPoster
 
         final SharedPreferences preferences = getActivity().getSharedPreferences(
@@ -238,14 +210,6 @@ public class DetailsFragment extends Fragment implements OnMapReadyCallback {
     }
 
     @Override
-    public void onResume() {
-        super.onResume();
-
-            mapFragment.getMapAsync(this);
-
-    }
-
-    @Override
     public void onDetach() {
         super.onDetach();
 
@@ -262,202 +226,5 @@ public class DetailsFragment extends Fragment implements OnMapReadyCallback {
         }
     }
 
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        myContext = (FragmentActivity) context;
-    }
 
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-
-        mMap = googleMap;
-
-        if(hasPermission());
-
-        else if (checkLocationEnabled())
-            isLocationEnabled = true;
-        else {
-            isLocationEnabled = false;
-            enableLocationDialog();
-        }
-
-        if (isPermissionGiven && isLocationEnabled) {
-            if (ActivityCompat.checkSelfPermission(getActivity(), android.Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(getActivity(), android.Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-                                //    ActivityCompat#requestPermissions
-                // here to request the missing permissions, and then overriding
-                //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-                //                                          int[] grantResults)
-                // to handle the case where the user grants the permission. See the documentation
-                // for ActivityCompat#requestPermissions for more details.
-                return;
-            }
-            mMap.setMyLocationEnabled(true);
-            mMap.setOnMyLocationButtonClickListener(
-                    new GoogleMap.OnMyLocationButtonClickListener() {
-                        @Override
-                        public boolean onMyLocationButtonClick() {
-
-                            Toast.makeText(getActivity(), "Loading...", Toast.LENGTH_SHORT).show();
-
-                            return false;
-                        }
-                    }
-            );
-        }
-
-
-        LatLng eventLocationLatLng = new LatLng(
-                Double.parseDouble(events.getLocationY()),
-                Double.parseDouble(events.getLocationX())
-        );
-
-        BitmapDrawable bitmapDrawable = (BitmapDrawable) getResources().getDrawable(R.drawable.festember_logo);
-        Bitmap bitmap =  bitmapDrawable.getBitmap();
-        bitmap = Bitmap.createScaledBitmap(bitmap, 80, 80, false);
-
-        mMap.addMarker(
-                new MarkerOptions()
-                        .position(
-                                eventLocationLatLng
-                        )
-                        .title(EventsAdapter.parseEventName(events.getName()) +
-                                " at " + events.getVenue())
-                        .icon(BitmapDescriptorFactory.fromBitmap(bitmap))
-        );
-
-        mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(
-                eventLocationLatLng,
-                16
-        ));
-
-
-    }
-
-    public boolean hasPermission(){
-        if (ActivityCompat.checkSelfPermission(getActivity(), android.Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(getActivity(), android.Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED)
-        {
-
-            isPermissionGiven = false;
-            isLocationEnabled = false;
-
-            ActivityCompat.requestPermissions(
-                    getActivity(),
-                    PERMISSIONS,
-                    PERMISSION_REQUEST_CODE
-            );
-
-            return false;
-        }
-
-        else
-            return true;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        switch (requestCode){
-
-            case PERMISSION_REQUEST_CODE:
-
-
-                if(grantResults.length==2){
-                    if(
-                            grantResults[0] == PackageManager.PERMISSION_GRANTED &&
-                                    grantResults[1] == PackageManager.PERMISSION_GRANTED
-                            )
-                        isPermissionGiven = true;
-                }
-                break;
-        }
-    }
-
-    public boolean checkLocationEnabled() {
-
-        LocationManager manager = (LocationManager) getActivity().getSystemService(Context.LOCATION_SERVICE);
-
-        return ((
-                manager.isProviderEnabled(LocationManager.GPS_PROVIDER)
-                        || manager.isProviderEnabled(LocationManager.NETWORK_PROVIDER)
-        ));
-
-    }
-    private void enableLocationDialog() {
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity())
-                .setTitle("Enable Location?")
-                .setMessage("You need to enable location to show your location on map")
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        Intent intent = new Intent(
-                                Settings.ACTION_LOCATION_SOURCE_SETTINGS
-                        );
-
-                        Log.d(LOG_TAG, "Starting settings implicit activity");
-
-                        startActivityForResult(intent, LOCATION_ENABLE_REQUEST_CODE);
-                    }
-                })
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        isLocationEnabled = false;
-                        dialog.dismiss();
-                    }
-                });
-
-        builder.show();
-
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        switch (requestCode) {
-
-            case LOCATION_ENABLE_REQUEST_CODE:
-
-                Log.d(LOG_TAG, "Inside onActivityResult");
-
-                if (checkLocationEnabled()) {
-                    isLocationEnabled = true;
-                    Log.d(LOG_TAG, "Location enabled");
-                } else {
-                    isLocationEnabled = false;
-                    Log.d(LOG_TAG, "Location not enabled");
-                }
-
-                if (
-                        ActivityCompat.checkSelfPermission(getActivity(),
-                                android.Manifest.permission.ACCESS_FINE_LOCATION) !=
-                                PackageManager.PERMISSION_GRANTED &&
-                                ActivityCompat.checkSelfPermission(getActivity(),
-                                        android.Manifest.permission.ACCESS_COARSE_LOCATION)
-                                        != PackageManager.PERMISSION_GRANTED) {
-                    isPermissionGiven = false;
-                    isLocationEnabled = false;
-                    return;
-                }
-
-                if (isPermissionGiven && isLocationEnabled && mMap != null)
-                {
-                    mMap.setMyLocationEnabled(true);
-                    mMap.setOnMyLocationButtonClickListener(
-                            new GoogleMap.OnMyLocationButtonClickListener() {
-                                @Override
-                                public boolean onMyLocationButtonClick() {
-
-                                    Toast.makeText(getActivity(), "Loading...", Toast.LENGTH_SHORT).show();
-
-                                    return false;
-                                }
-                            }
-                    );
-                }
-
-
-                break;
-
-        }
-    }
 }
