+++ app/src/main/java/com/festember16/app/MyPagerAdapter.java
@@ -34,6 +34,10 @@ public class MyPagerAdapter extends FragmentPagerAdapter {
                 RulesFragment rulesFragment = RulesFragment.getInstance(eventId);
                 return rulesFragment;
 
+            case 3:
+                MapsTabFragment mapsTabFragment = new MapsTabFragment();
+                return mapsTabFragment;
+
             default:
                 DetailsFragment detailsFragment = DetailsFragment.getInstance(eventId);
                 return detailsFragment;
@@ -43,7 +47,7 @@ public class MyPagerAdapter extends FragmentPagerAdapter {
 
     @Override
     public int getCount() {
-        return 3;
+        return 4;
     }
 
     @Override
@@ -56,6 +60,8 @@ public class MyPagerAdapter extends FragmentPagerAdapter {
 
             case 2:
                 return "Rules";
+
+            case 3: return "Map";
 
             default:
                 return "Details";
