+++ app/src/main/res/layout/details_layout.xml
@@ -27,20 +27,6 @@
         android:textStyle="bold|italic"
         android:typeface="serif" />
 
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/mapContainer">
-
-        <fragment
-            android:layout_width="match_parent"
-            android:layout_height="180dp"
-            android:id="@+id/mapPreview"
-            android:layout_marginTop="16dp"
-            class="com.google.android.gms.maps.SupportMapFragment" />
-    </LinearLayout>
-
     <Space
         android:layout_width="20px"
         android:layout_height="20px" />
@@ -91,36 +77,38 @@
                 android:layout_width="20px"
                 android:layout_height="20px" />
 
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textAppearance="?android:attr/textAppearanceLarge"
-                android:id="@+id/eventCluster"
-                android:layout_below="@+id/eventStartTime"
-                android:layout_centerHorizontal="true"
-                android:text="Cluster"
-                android:textColor="@color/textColorLight"
-                android:textStyle="bold|italic"
-                android:typeface="serif" />
+            <!--<TextView-->
+                <!--android:layout_width="wrap_content"-->
+                <!--android:layout_height="wrap_content"-->
+                <!--android:textAppearance="?android:attr/textAppearanceLarge"-->
+                <!--android:id="@+id/eventCluster"-->
+                <!--android:layout_below="@+id/eventStartTime"-->
+                <!--android:layout_centerHorizontal="true"-->
+                <!--android:text="Cluster"-->
+                <!--android:textColor="@color/textColorLight"-->
+                <!--android:textStyle="bold|italic"-->
+                <!--android:typeface="serif" -->
+                <!--android:visibility="gone"/>-->
 
             <Space
                 android:layout_width="20px"
                 android:layout_height="20px" />
 
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Register"
-                android:id="@+id/registerButton"
-                android:background="@color/buttonBackground"
-                android:textColor="@color/textColorLight"
-                android:textSize="24sp"
-                android:textStyle="bold"
-                android:padding="10dp"
-                android:layout_marginTop="20dp"
-                android:typeface="serif" />
 
         </LinearLayout>
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Register"
+            android:id="@+id/registerButton"
+            android:background="@color/buttonBackground"
+            android:textColor="@color/textColorLight"
+            android:textSize="24sp"
+            android:textStyle="bold"
+            android:padding="10dp"
+            android:layout_marginTop="20dp"
+            android:typeface="serif" />
 
     </ScrollView>
